Fix concurrency issues
